Allow for fractional velocities

+++ src/ecs/components/position.cpp
@@ -1,25 +1,25 @@
 #include "position.h"
 
 
-Position::Position(int x, int y) {
+Position::Position(double x, double y) {
     m_x = x;
     m_y = y;
 }
 
-int Position::get_x() {
+double Position::get_x() {
     return m_x;
 }
 
-int Position::get_y() {
+double Position::get_y() {
     return m_y;
 }
 
 
-void Position::set_x(int val) {
+void Position::set_x(double val) {
     m_x = val;
 }
 
 
-void Position::set_y(int val) {
+void Position::set_y(double val) {
     m_y = val;
 }
